Crypto-Pipeline/alerting/requirements.txt
# Alert Consumer Python Dependencies
# Note: The Dockerfile installs confluent-kafka separately with system libs.

# Load environment variables from .env
python-dotenv==1.0.0
